requirements.txt
streamlit>=1.30.0
moviepy
google-generativeai>=0.3.2
openai>=1.3.7
ffmpeg-python
pydub

